feat: 단위 변환 long

+++ src/main/java/lotto/model/LottoCount.java
@@ -5,18 +5,18 @@ import static lotto.utils.ExceptionMessage.AMOUNT_NOT_DIVISIBLE_EXCEPTION;
 public class LottoCount {
     private static final int LOTTO_PRICE = 1000;
 
-    private final int amount;
+    private final long amount;
 
-    public LottoCount(int amount) {
+    public LottoCount(long amount) {
         validateAmount(amount);
         this.amount = amount;
     }
 
-    public int calculateLottoCount(){
+    public long calculateLottoCount(){
         return amount / LOTTO_PRICE;
     }
 
-    private void validateAmount(int amount){
+    private void validateAmount(long amount){
         if(amount % LOTTO_PRICE !=0){
             throw new IllegalArgumentException(AMOUNT_NOT_DIVISIBLE_EXCEPTION);
         }
